site_app/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
Django==5.1.1
django-allauth==64.2.1
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
filelock==3.16.0
fsspec==2024.9.0
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
httplib2==0.22.0
huggingface-hub==0.24.7
idna==3.8
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
oauthlib==3.2.2
packaging==24.1
proto-plus==1.24.0
protobuf==5.28.1
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
sqlparse==0.5.1
sympy==1.13.2
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
